+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 Unreleased in the current development version (target v0.24.0):
 
 Complete list:
+- Add tests for diagnostics CLI (#219)
 - Expose `__version__` from `aqua.core.version` module (#216)
 - ECmean: time selection is now allowed (#178)